backend/requirements.txt
fastapi
uvicorn[standard]
motor>=3.7,<4
pymongo>=4.9,<5
beanie
pydantic-settings
paho-mqtt
apscheduler
python-dotenv
